🧬 Migrate the base data modal *_HasItemsPropConfig* into base module *_base*.
